move useNuxSwipe to proper folder
